chg: slight refactor, moved logging into separate modules
