+++ generator_utils.py
@@ -1,0 +1,93 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# NOTE: THIS IS THE CLASS THAT WILL STORE ALL OF THE EXCEPTIONS ETC
+
+"""Scarlett Generator Object audio utils"""
+
+from __future__ import with_statement
+from __future__ import division
+
+import sys
+import os
+
+import argparse
+import pprint
+pp = pprint.PrettyPrinter(indent=4)
+
+try:
+    import queue
+except ImportError:
+    import Queue as queue
+
+try:
+    from urllib.parse import quote
+except ImportError:
+    from urllib import quote
+
+
+QUEUE_SIZE = 10
+BUFFER_SIZE = 10
+SENTINEL = '__GSTDEC_SENTINEL__'
+
+import signal
+
+from IPython.core.debugger import Tracer
+from IPython.core import ultratb
+
+sys.excepthook = ultratb.FormattedTB(mode='Verbose',
+                                     color_scheme='Linux',
+                                     call_pdb=True,
+                                     ostream=sys.__stdout__)
+import generator_log
+
+
+class DecodeError(Exception):
+    """The base exception class for all decoding errors raised by this
+    package.
+    """
+
+
+class NoBackendError(DecodeError):
+    """The file could not be decoded by any backend. Either no backends
+    are available or each available backend failed to decode the file.
+    """
+
+
+def _gst_available():
+    """Determine whether Gstreamer and the Python GObject bindings are
+    installed.
+    """
+    try:
+        import gi
+    except ImportError:
+        return False
+
+    try:
+        gi.require_version('Gst', '1.0')
+    except (ValueError, AttributeError):
+        return False
+
+    try:
+        from gi.repository import Gst  # noqa
+    except ImportError:
+        return False
+
+    return True
+
+
+def audio_open(path):
+    """Open an audio file using a library that is available on this
+    system.
+    """
+    # GStreamer.
+    if _gst_available():
+        from . import generator_player
+        try:
+            return generator_player.ScarlettPlayer(path)
+            # return gstdec.ScarlettPlayer(path)
+        except DecodeError:
+            pass
+
+    # All backends failed!
+    raise NoBackendError()
